[Refactor] gpt client에 집중되어 있던 메서드를 extractor, interviewer로 분리
